Added starter files/ added show page funtion
I accidently created this repo after I started the project.  I added all starter files to the repo and as well as the showPage function that checks if the index of the array Is between 0 and 9, if that is the case it creates multiple elements and adds them to the page to show student data.