updated variables to "const", added "addPagination" fuction
I have updated my variables from "let" to "const" as these variables will not be changed.  I also added a function that checks how many pages are needed based on the student data provided and adds the appropriate  number of pages to display 9 students each page, with the remainder on the last.  The function also includes an event listener that listens for clicks on the pagination buttons, removes the active class from the previous button and assigns it to the new event.